+++ packages/eslint-plugin/tests/rules/no-unnecessary-type-assertion.test.ts
@@ -202,6 +202,12 @@ let a: { b?: string } | undefined;
 a!.b = '';
       `,
     },
+    `
+let value: number | undefined;
+let values: number[] = [];
+
+value = values.pop()!;
+    `,
   ],
 
   invalid: [
@@ -358,6 +364,22 @@ function foo<T extends string>(bar: T) {
     },
     {
       code: `
+declare const foo: Foo;
+const bar = <Foo>foo;
+      `,
+      output: `
+declare const foo: Foo;
+const bar = foo;
+      `,
+      errors: [
+        {
+          messageId: 'unnecessaryAssertion',
+          line: 3,
+        },
+      ],
+    },
+    {
+      code: `
 declare function nonNull(s: string | null);
 let s: string | null = null;
 nonNull(s!);
@@ -477,14 +499,10 @@ y! = 0;
       output: `
 let x: number | undefined;
 let y: number | undefined;
-y = x;
+y = x!;
 y = 0;
       `,
       errors: [
-        {
-          messageId: 'contextuallyUnnecessary',
-          line: 4,
-        },
         {
           messageId: 'contextuallyUnnecessary',
           line: 5,