+++ packages/eslint-plugin/tests/rules/no-unsafe-return.test.ts
@@ -9,7 +9,7 @@ import {
 const ruleTester = new RuleTester({
   parser: '@typescript-eslint/parser',
   parserOptions: {
-    project: './tsconfig.json',
+    project: './tsconfig.noImplicitThis.json',
     tsconfigRootDir: getFixturesRootDir(),
   },
 });
@@ -290,6 +290,31 @@ receiver(function test() {
             sender: 'Set<any>',
             receiver: 'Set<string>',
           },
+        },
+      ],
+    },
+    {
+      code: `
+function foo() {
+  return this;
+}
+
+function bar() {
+  return () => this;
+}
+      `,
+      errors: [
+        {
+          messageId: 'unsafeReturnThis',
+          line: 3,
+          column: 3,
+          endColumn: 15,
+        },
+        {
+          messageId: 'unsafeReturnThis',
+          line: 7,
+          column: 16,
+          endColumn: 20,
         },
       ],
     },